requirements.txt
altair==4.1.0
flask==1.1.2
gunicorn==20.0.4
pandas==1.2.4
psycopg2==2.8.6
python-dotenv==0.17.1
sqlalchemy==1.4.11
vega_datasets==0.8.0
requests==2.25.1
jinja2==3.0.1
numpy==1.19.4